Proyecto/Software/requirements.txt
# Core dependencies for EEG-BCI Motor Imagery Classification
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# EEG processing
mne>=1.5.0
pyedflib>=0.1.30

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Deep Learning (optional - uncomment if needed)
# tensorflow>=2.13.0
# torch>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
